src/requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
awesome-slugify==1.6.5
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.7
django-behaviors==0.4.1
django-bootstrap-form==3.4
django-debug-toolbar==1.9.1
django-environ==0.4.5
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.8.2
djangorestframework-filters==0.10.2
Faker==0.8.8
flake8==3.5.0
flake8-isort==2.5
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
mccabe==0.6.1
mixer==6.0.1
more-itertools==4.2.0
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.5.4
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.6.3
pytest-django==3.3.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
regex==2018.6.21
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
testfixtures==6.2.0
text-unidecode==1.2
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.23
wcwidth==0.1.7
